Standardize rollup config.
