+++ packages/gensx-anthropic/package.json
@@ -52,13 +52,9 @@
     "@anthropic-ai/sdk": "^0.39.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "catalog:",
-    "@rollup/plugin-node-resolve": "catalog:",
-    "@rollup/plugin-typescript": "catalog:",
     "@types/node": "catalog:packages",
     "@vitest/coverage-istanbul": "catalog:",
     "@anthropic-ai/sdk": "^0.39.0",
-    "rollup": "catalog:",
     "vitest": "catalog:"
   },
   "publishConfig": {
